+++ email/src/backend/maildir/config.rs
@@ -1,13 +1,18 @@
-//! Maildir backend config module.
+//! Module dedicated to the Maildir backend configuration.
 //!
-//! This module contains the representation of the Maildir backend
-//! configuration of the user account.
+//! This module contains the configuration specific to the Maildir
+//! backend.
 
 use std::path::PathBuf;
 
-/// Represents the Maildir backend config.
+/// The Maildir backend configuration.
 #[derive(Debug, Default, Clone, Eq, PartialEq)]
 pub struct MaildirConfig {
-    /// Represents the Maildir root directory.
+    /// The Maildir root directory.
+    ///
+    /// The path should point to the root level of the Maildir
+    /// directory (the one containing the `cur`, `new` and `tmp`
+    /// folders). Path is shell-expanded, which means environment
+    /// variables and tilde `~` are replaced by their values.
     pub root_dir: PathBuf,
 }